perf: lazy load design-linking feature to reduce main bundle size
- Lazy load DesignLinkingDialogs in App.tsx and MobileLayout.tsx
- Lazy load LinkedDesignSection in SingleBinInspector
- Extract BinContextMenuDesignSection for lazy loading hooks/UI
- Bump size limits: main 168→170 kB, total 965→1000 kB

Reduces main bundle impact from ~11 kB to ~1.25 kB (88% reduction).
Design-linking code now loads on-demand when bin_designer flag enabled.
